refactor: create content dialog
