+++ frontend/src/components/chat/dialog/source-dialog.tsx
@@ -1,18 +1,7 @@
-import CloseIcon from '@mui/icons-material/Close';
-import Dialog from '@mui/material/Dialog';
-import DialogContent from '@mui/material/DialogContent';
-import DialogTitle from '@mui/material/DialogTitle';
-import IconButton from '@mui/material/IconButton';
+import ContentDialog from './content-dialog';
 import Typography from '@mui/material/Typography';
-import { styled } from '@mui/material/styles';
 
-const BootstrapDialog = styled(Dialog)(({ theme }) => ({
-  '& .MuiDialogContent-root': {
-    padding: theme.spacing(4),
-  },
-}));
-
-interface ContentDialogProps {
+interface SourceDialogProps {
   source: string;
   showDialog: boolean;
   closeDialog: () => void;
@@ -22,33 +11,11 @@ const SourceDialog = ({
   source,
   showDialog,
   closeDialog,
-}: ContentDialogProps) => {
+}: SourceDialogProps) => {
   return (
-    <BootstrapDialog
-      onClose={closeDialog}
-      aria-labelledby='customized-dialog-title'
-      open={showDialog}
-      maxWidth={'lg'}
-    >
-      <DialogTitle sx={{ m: 0, p: 2 }} id='customized-dialog-title'>
-        Source
-      </DialogTitle>
-      <IconButton
-        aria-label='close'
-        onClick={closeDialog}
-        sx={(theme) => ({
-          position: 'absolute',
-          right: 8,
-          top: 8,
-          color: theme.palette.grey[500],
-        })}
-      >
-        <CloseIcon />
-      </IconButton>
-      <DialogContent sx={{ alignItems: 'center' }}>
-        <Typography style={{ whiteSpace: 'pre-wrap' }}>{source}</Typography>
-      </DialogContent>
-    </BootstrapDialog>
+    <ContentDialog showDialog={showDialog} closeDialog={closeDialog}>
+      <Typography style={{ whiteSpace: 'pre-wrap' }}>{source}</Typography>
+    </ContentDialog>
   );
 };
